Implementation and structuring of the header section
Initial use of the 'require' statement selects the content of 'head.php'
and transfers it into this class.

Creation of the banner then follows; with the assignment of its location
'col-md-4' and the formation of the Title of my site within the banner
as 'Red Carpet Films'.

A concluding require statement is made in relation to the
'navigation.php' class; to initiate its link within the class.