Construction of the Navigation Bar; allowing users access to the Films within the site
To begin with two require statements are issued; to the 'db.php' and
'genre.class.php' respectively.

The implementation of the Navigation feature then follows.

Links are then established with the pages of the site; whilst, a drop
down menu is created as an enhancement of interactivity within the site.

After implementing the drop down menu; a query is ran to the database.
This query is initiated to 'Collect all data from the
'comp2203-cw-1415.genre' table; and order them by name. A while loop is
also used to help fetch the genres from the relevant column in the
database.