[2] Introduction message to the site and use of queries for the selection of random films from the database
An introduction message was then written and implemented onto the
homepage to welcome all users to the site; whilst, also giving them a
flavour of what to expect within the sites's contents.

A random film generator was then implemented through the use of queries;
to ensure a new film is present on the homepage every time it is
reloaded.