[1] Creation of the film class; information about each film is displayed using the GET parameter
Numerous require statements are initially stated for classes.

Content on the film page is initiated through the creation of a
container. A header is then established; containing an includes
statement for the 'header.php' class.

The use of a GET parameter is then used within an If statement to
information about films from the database.