Creation of the Role class; extracts information on the actors name, film and the character they played from the database
Variables of ID, Actor ID , Film ID and character are all extrapolated
for the role table.

A function is created to get actos; and a require statement containing
an include statement for the 'db.php' class is stated.

A mysql query is then ran to extract actor information from the
database.

The query is simplified to 'Selecting all from the 'comp2203-cw-1415'
table where the ID variable is equal to the actor id.

A use of a while loop assists in fetching the actor information from the
actor table.