[1] Formation of the review submit class for users of the site
Require statements are administered for the 'db.php', 'header.php',
'review.class.php' and 'fil.class.php' classes.

The use of the try-catch block helps to validate the submission of
reviews. The input of invalid names and comments through a numbers for
example; results in an exception with the appropriate error message
being thrown.