[2] Allowing users to review, comment and like any film
Through the use of the POST variable; users are given authority to
review, comment and like any film they like via a radio button.

The implementation of the personable review system; updates the
specified fields within the back end of the database.

The implementation of a review confirmation message for successful
submits is implemented. It reads ' Thank you for submitting a review';
this acknowledges the success of the review to the customer.