[3] Use of exceptions to display error messages and show confirmation
Exception is caught successfully and error message is displayed if
reviews aren't submitted successfully.

If review isn't submitted successfully; error message saying 'Your
review was not submitted. A further message 'Please enter your name and
a valid comment' is issued. An option is also administered reading 'Back
to view Films to rent' to any customers wanting to redirect their route.

This signals the end of the content; a require statement is put into the
footer which contains an include of the 'footer.php' class.