[1] Creation of the genre class; activating the search by genre feature located within the navigation bar
Implementation of initial require statements are first established; the
'db.php' and 'film.class.php' classes are connected too.

The genre page is put into a container; and a require header contains an
inclusive 'header.php' class.

Utilisation of the IF statement  and inclusive GET id variable is used
to identify the films searched by genre within the navigation bar.